+++ react/StoreContextProvider.js
@@ -1,10 +1,10 @@
 import React, { Component } from 'react'
-import { Helmet } from 'render'
+import { Helmet, withRuntimeContext } from 'render'
 import PropTypes from 'prop-types'
 
+import GtmScripts from './components/GtmScripts'
 import { OrderFormProvider } from './OrderFormContext'
 import { DataLayerProvider } from './components/withDataLayer'
-import { gtmScript, gtmFrame } from './scripts/gtm'
 
 const APP_LOCATOR = 'vtex.store'
 const CONTENT_TYPE = 'text/html;charset=utf-8'
@@ -36,6 +36,25 @@ class StoreContextProvider extends Component {
     }
   }
 
+  sendPageViewEvent = () => {
+    this.pushToDataLayer({
+      event: 'pageView',
+      pageTitle: document.title,
+      pageUrl: location.href,
+      accountName: this.props.runtime.account,
+    })
+  }
+
+  componentDidMount() {
+    this.sendPageViewEvent()
+  }
+
+  componentDidUpdate(prevProps) {
+    if (prevProps.runtime.route.path !== this.props.runtime.route.path) {
+      this.sendPageViewEvent()
+    }
+  }
+
   render() {
     const { country, locale, currency } = global.__RUNTIME__.culture
     const settings = this.context.getSettings(APP_LOCATOR) || {}
@@ -47,11 +66,6 @@ class StoreContextProvider extends Component {
       metaTagRobots,
       storeName,
     } = settings
-    const scripts = gtmId ? [{
-      'type': 'application/javascript',
-      'innerHTML': gtmScript(gtmId),
-    }] : []
-    const noscripts = gtmId ? [{ id: 'gtm_frame', innerHTML: gtmFrame(gtmId) }] : []
 
     this.initDataLayer()
     return (
@@ -61,7 +75,7 @@ class StoreContextProvider extends Component {
           set: this.pushToDataLayer,
         }}
       >
-        <Helmet script={scripts} noscript={noscripts} />
+        <GtmScripts gtmId={gtmId} />
         <Helmet>
           <title>{titleTag}</title>
           <meta name="description" content={metaTagDescription} />
@@ -86,4 +100,4 @@ StoreContextProvider.contextTypes = {
   getSettings: PropTypes.func,
 }
 
-export default StoreContextProvider
+export default withRuntimeContext(StoreContextProvider)